move k8s diagnostics to dedicated file
